+++ packages/moonstone/Scrollable/ScrollableNative.js
@@ -3,6 +3,7 @@ import {constants, ScrollableBaseNative as UiScrollableBaseNative} from '@enact/
 import {getDirection} from '@enact/spotlight';
 import {getTargetByDirectionFromElement, getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
 import {Job} from '@enact/core/util';
+import platform from '@enact/core/platform';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import ReactDOM from 'react-dom';
@@ -216,6 +217,9 @@ class ScrollableBaseNative extends Component {
 
 	// browser native scrolling
 	resetPosition = null // prevent auto-scroll on focus by Spotlight
+
+	isVoiceControl = false
+	voiceControlDirection = 'vertical'
 
 	onMouseDown = () => {
 		this.lastFocusedItem = null;
@@ -544,6 +548,10 @@ class ScrollableBaseNative extends Component {
 		this.lastFocusedItem = null;
 		this.lastScrollPositionOnFocus = null;
 		this.isWheeling = false;
+		if (this.isVoiceControl) {
+			this.isVoiceControl = false;
+			this.updateFocusAfterVoiceControl();
+		}
 	}
 
 	focusOnItem () {
@@ -655,6 +663,10 @@ class ScrollableBaseNative extends Component {
 			childContainerRef.addEventListener('mouseover', this.onMouseOver, {capture: true});
 			childContainerRef.addEventListener('mousemove', this.onMouseMove, {capture: true});
 			childContainerRef.addEventListener('focusin', this.onFocus);
+			if (platform.webos) {
+				childContainerRef.addEventListener('webOSVoice', this.onVoice);
+				childContainerRef.setAttribute('data-webos-voice-intent', 'Scroll');
+			}
 		}
 	}
 
@@ -664,6 +676,10 @@ class ScrollableBaseNative extends Component {
 			childContainerRef.removeEventListener('mouseover', this.onMouseOver, {capture: true});
 			childContainerRef.removeEventListener('mousemove', this.onMouseMove, {capture: true});
 			childContainerRef.removeEventListener('focusin', this.onFocus);
+			if (platform.webos) {
+				childContainerRef.removeEventListener('webOSVoice', this.onVoice);
+				childContainerRef.removeAttribute('data-webos-voice-intent');
+			}
 		}
 	}
 
@@ -693,6 +709,71 @@ class ScrollableBaseNative extends Component {
 		if (ref) {
 			this.uiRef = ref;
 		}
+	}
+
+	updateFocusAfterVoiceControl = () => {
+		const spotItem = Spotlight.getCurrent();
+		if (spotItem && this.uiRef.containerRef.contains(spotItem)) {
+			const
+				viewportBounds = this.uiRef.containerRef.getBoundingClientRect(),
+				spotItemBounds = spotItem.getBoundingClientRect(),
+				nodes = Spotlight.getSpottableDescendants(this.uiRef.containerRef.dataset.spotlightId),
+				first = this.voiceControlDirection === 'vertical' ? 'top' : 'left',
+				last = this.voiceControlDirection === 'vertical' ? 'bottom' : 'right';
+
+			if (spotItemBounds[last] < viewportBounds[first] || spotItemBounds[first] > viewportBounds[last]) {
+				for (let i = 0; i < nodes.length; i++) {
+					const nodeBounds = nodes[i].getBoundingClientRect();
+					if (nodeBounds[first] > viewportBounds[first] && nodeBounds[last] < viewportBounds[last]) {
+						Spotlight.focus(nodes[i]);
+						break;
+					}
+				}
+			}
+		}
+	}
+
+	onVoice = (e) => {
+		const scroll = e && e.detail && e.detail.scroll;
+		this.isVoiceControl = true;
+
+		switch (scroll) {
+			case 'up':
+				this.voiceControlDirection = 'vertical';
+				this.onScrollbarButtonClick({isPreviousScrollButton: true, isVerticalScrollBar: true});
+				break;
+			case 'down':
+				this.voiceControlDirection = 'vertical';
+				this.onScrollbarButtonClick({isPreviousScrollButton: false, isVerticalScrollBar: true});
+				break;
+			case 'left':
+				this.voiceControlDirection = 'horizontal';
+				this.onScrollbarButtonClick({isPreviousScrollButton: true, isVerticalScrollBar: false});
+				break;
+			case 'right':
+				this.voiceControlDirection = 'horizontal';
+				this.onScrollbarButtonClick({isPreviousScrollButton: false, isVerticalScrollBar: false});
+				break;
+			case 'top':
+				this.voiceControlDirection = 'vertical';
+				this.uiRef.scrollTo({align: 'top'});
+				break;
+			case 'bottom':
+				this.voiceControlDirection = 'vertical';
+				this.uiRef.scrollTo({align: 'bottom'});
+				break;
+			case 'leftmost':
+				this.voiceControlDirection = 'horizontal';
+				this.uiRef.scrollTo({align: 'left'});
+				break;
+			case 'rightmost':
+				this.voiceControlDirection = 'horizontal';
+				this.uiRef.scrollTo({align: 'right'});
+				break;
+			default:
+				this.isVoiceControl = false;
+		}
+		e.preventDefault();
 	}
 
 	render () {