+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `moonstone/TooltipDecorator` to prevent unnecessary re-renders when losing focus
+
 ## [2.0.0-beta.8] - 2018-06-25
 
 ### Added